+++ src/main/kotlin/trans/rights/client/manager/impl/CommandManager.kt
@@ -8,7 +8,7 @@ import trans.rights.client.modules.command.impl.*
 
 object CommandManager : Manager<Command>(mutableSetOf()) {
     override fun load() {
-        this.add(listOf(HackCommand, ReloadCommand))
+        this.add(listOf(CHelpCommand, HackCommand, ReloadCommand))
     }
 
     @JvmStatic
