+++ src/main/kotlin/trans/rights/client/modules/command/impl/CHelpCommand.kt
@@ -5,7 +5,7 @@ import net.minecraft.server.command.CommandManager.*
 import net.minecraft.server.command.ServerCommandSource
 import trans.rights.client.modules.command.Command
 
-object HelpCommand : Command("help", "Informs you about the different features of this client") {
+object CHelpCommand : Command("chelp", "Informs you about the different features of this client") {
     override fun register(dispatcher: CommandDispatcher<ServerCommandSource>) {
         dispatcher.register(literal("/$name").executes { helpMessage() })
     }
