+++ src/main/kotlin/trans/rights/client/modules/hack/Hack.kt
@@ -5,11 +5,11 @@ import com.google.gson.JsonPrimitive
 import java.io.File
 import trans.rights.client.TransRights.Companion.LOGGER
 import trans.rights.client.manager.impl.HackManager
-import trans.rights.client.manager.impl.Settings
-import trans.rights.client.misc.api.Globals
+import trans.rights.client.modules.setting.Settings
 import trans.rights.client.modules.Module
-import trans.rights.client.modules.setting.settings.BooleanSetting
-import trans.rights.client.modules.setting.settings.NumberSetting
+import trans.rights.client.modules.setting.impl.BooleanSetting
+import trans.rights.client.modules.setting.impl.EnumSetting
+import trans.rights.client.modules.setting.impl.NumberSetting
 import trans.rights.client.util.file.FileHelper
 import trans.rights.event.bus.impl.BasicEventManager
 
@@ -19,7 +19,7 @@ abstract class Hack(
     val settings: Settings = Settings(),
     val file: File = File(HackManager.directory.absolutePath + "$name.json"),
     private var enabled: Boolean = false
-) : Module(name, description), Globals {
+) : Module(name, description) {
 
     fun enable() {
         BasicEventManager.register(this)
@@ -41,6 +41,7 @@ abstract class Hack(
 
     open fun onDisable() {}
 
+    // TODO: Make this better / more streamlined
     fun load(file: File) {
         try {
             if (!file.exists()) file.createNewFile()
@@ -74,7 +75,11 @@ abstract class Hack(
             val json = JsonObject()
 
             for (setting in settings.values) {
-                json.add(setting.name, JsonPrimitive(setting.value.toString()))
+                when (setting) {
+                    is BooleanSetting -> json.add(setting.name, JsonPrimitive(setting.value))
+                    is NumberSetting -> json.add(setting.name, JsonPrimitive(setting.value))
+                    is EnumSetting -> json.add(setting.name, JsonPrimitive(setting.value.toString()))
+                }
             }
 
             FileHelper.writeToJson(json, file.toPath())
