+++ src/main/kotlin/trans/rights/client/modules/hack/impl/AutoHit.kt
@@ -1,13 +1,14 @@
 package trans.rights.client.modules.hack.impl
 
 import trans.rights.client.events.TickEvent
+import trans.rights.client.misc.api.Globals
 import trans.rights.client.modules.hack.Hack
-import trans.rights.client.modules.setting.settings.BooleanSetting
-import trans.rights.client.modules.setting.settings.NumberSetting
+import trans.rights.client.modules.setting.impl.BooleanSetting
+import trans.rights.client.modules.setting.impl.NumberSetting
 import trans.rights.event.annotation.EventHandler
 import trans.rights.event.listener.impl.*
 
-object AutoHit : Hack("Auto-hit", "Automatically hit people near you") {
+object AutoHit : Hack("Auto-hit", "Automatically hit people near you"), Globals {
     private val waitForDelay = BooleanSetting("Wait", true)
     private val swap = BooleanSetting("Swap", true)
     private val delay = NumberSetting("tick-delay", 6)
