+++ src/main/kotlin/trans/rights/client/modules/hack/impl/FlightHack.kt
@@ -9,8 +9,8 @@ import trans.rights.client.events.TickEvent
 import trans.rights.client.misc.api.Globals
 import trans.rights.client.misc.api.Globals.mc
 import trans.rights.client.modules.hack.Hack
-import trans.rights.client.modules.setting.settings.BooleanSetting
-import trans.rights.client.modules.setting.settings.NumberSetting
+import trans.rights.client.modules.setting.impl.BooleanSetting
+import trans.rights.client.modules.setting.impl.NumberSetting
 import trans.rights.event.annotation.EventHandler
 import trans.rights.event.listener.impl.*
 
@@ -37,6 +37,8 @@ object FlightHack : Hack("Flight", "Fly using hacks"), Globals {
     init {
         settings.add(speed)
         settings.add(cancelSpeed)
+
+        this.enable()
     }
 
     override fun onEnable() {
